Updates
- Update ffmpeg from 8.0 to 8.1
- Update dav1d to from 1.5.1 to 1.5.3
- Update libde265 from 1.0.16 to 1.0.18
- Update libheif from 1.20.1 to 1.21.2
- Update ImageMagick from 7.1.2-0 to 7.1.2-18
- Update graphicsmagick from 1.3.45 to 1.3.46
- Update frei0r from 2.3.3 to 2.5.6
- Add `LANG` env variable
- Update coffescript/node/typescript dependencies
- Update prolog from 9.3.20 to 9.3.26
- Update python dependencies

- Add network pruning after every execution.

- Update default stdin limit from 8mb to 16mb

+++ notsocode_server/app.py
@@ -36,8 +36,9 @@ def before_request(request: Request):
 
 @app.on_response
 async def after_request(request: Request, response: HTTPResponse):
-    for job in request.ctx.jobs:
-        await job.kill()
+    if hasattr(request.ctx, 'jobs'):
+        for job in request.ctx.jobs:
+            await job.kill()
 
     if hasattr(request.ctx, 'started'):
         started = request.ctx.started
